refactor(instance-ai): split tracing payload helpers
